Code refactoring and reusability
