+++ src/components/index.ts
@@ -2,10 +2,21 @@ import Button from "./Button";
 import Block from "./Block";
 import PurchaseComponent from "./PurchaseComponent";
 import SelectableMenu from "./SelectableMenu";
+import Collection from "./Collections/Collection";
+import Collections from "./Collections/Collections";
+
+import NFTTokens from "./NFTTokens/NFTTokens";
+import Token from "./NFTTokens/Token";
+import Offer from "./NFTTokens/Offer";
 
 export {
   Button,
   Block,
   PurchaseComponent,
-  SelectableMenu
+  SelectableMenu,
+  Collections,
+  Collection,
+  NFTTokens,
+  Token,
+  Offer
 }
